fix(services): resolve remaining bugs from bug_hunt_v3_report
- F1 A2AAdapter: include timeout failure prefix in HTTPStatusError
- F2 CLIAdapter: avoid infinite block on proc.communicate with timeout=None
- F5 MCPAdapter: do not mistakenly mark 65535 valid ports as unreachable
- F2 MiddlewareChain: correctly flag misconfigured 400 responses
- Fix tests/test_coverage_supplementary*.py to properly pass role='admin' to api actions

+++ 03-development/src/app/services/ab_testing.py
@@ -173,7 +173,7 @@ class ABTestManager:
         for variant, weight in traffic_split.items():
             if not isinstance(weight, (int, float)) or weight < 0:
                 continue
-            cumulative += int(weight)
+            cumulative += float(weight)
             if bucket < cumulative:
                 return str(variant)
         return ABTestManager._CONTROL_FALLBACK
